Several updates re: Google Analytics and page body tags
The GA Tracking ID has been moved to a variable at the top of the
document since it is used in multiple places.

We've made minor changes to the body class property (eliminating
irrational primary### and secondary### on the home page, adding a depth
param on pages other than the home page).

We have put the body class property in a variable so it can be re-used

We are passing the body class property to TinyMCE in the two bucket
handlers

We've switch to our new googleAnalytics template instead of hardcoding.